working on #13 fixes null pointer exception in loadFileData()
realized casinoInfo on ln74 was null because it was referencing the
unassigned file object in the class's fields. changed method to accept
File as an input parameter to make sure a file is passed into the method
before it runs. also changed fileSetup() ln60 to assign the FILE_PATH
field to the empty casinoInfo object, so that the file object is loaded
throughout the program