working on #13 findTopPlayers()
previous attempt was throwing ConcurrentModificationException error,
learned through this post
https://stackoverflow.com/questions/26418325/java-exception-in-thread-main-java-util-concurrentmodificationexception
that we cannot modify it the array list topPlayers while iterating using
the enhanced for loop; on line225,changed to a regular for loop
iterating by index. adjusted rest of method fields to accommodate this
change; program now gets to the point of accepting user input 't'
successfully without showing an error. next,we'll move on to working on
top players report display