working on top player report
report is now displaying when selected from menu.
GameManager.findTopPlayers() sorts top players but doesn't remove excess
player records from array. It currently returns a list that repeats
itself in an as-yet-undetermined pattern. However, this would be
functional-for-now if we only *display* the indexes of list items for
the number of top players we wish to display, or trim the topPlayers
ArrayList to the number of topPlayers before returning.